requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
numpy==1.26.4
scipy==1.13.1
tqdm==4.66.4
python-dotenv==1.0.1

# ── STT Server (VPS Contabo) ──────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
faster-whisper==1.0.3

# ── Voice / Wake-word (Laptop) ────────────────────────────────────────────────
openwakeword==0.6.0
pyaudio==0.2.14
sounddevice==0.4.7
soundfile==0.12.1
websockets==12.0

# ── Emotion / HuggingFace (MLOps demo) ───────────────────────────────────────
datasets==2.20.0

# ── Vision / CIFAR-10 (MLOps demo) ───────────────────────────────────────────
torch==2.3.1
torchvision==0.18.1
matplotlib==3.9.0
scikit-learn==1.5.0
Pillow==10.3.0

# ── Reporting / CML ───────────────────────────────────────────────────────────
pandas==2.2.2
tabulate==0.9.0

# ── LLM / QLoRA (Laptop inference) ───────────────────────────────────────────
peft>=0.11.0
transformers>=4.41.0
accelerate>=0.30.0

# ── TTS ───────────────────────────────────────────────────────────────────────
edge-tts>=6.1.9
pygame>=2.5.2
